tweak for 64bit and use of NSUInteger
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/sqlclient/trunk@34445 72102866-910b-0410-8b05-ffd578937521

+++ SQLClient.h
@@ -267,7 +267,7 @@ extern NSString * const SQLClientDidDisconnectNotification;
 /** <override-subclass />
  * Returns the object at the specified indes.<br />
  */
-- (id) objectAtIndex: (unsigned)index;
+- (id) objectAtIndex: (NSUInteger)index;
 
 /**
  * Returns the value of the named field.<br />
@@ -279,7 +279,7 @@ extern NSString * const SQLClientDidDisconnectNotification;
  * Replaces the value at the specified index.<br />
  * Subclasses must implement this method.
  */
-- (void) replaceObjectAtIndex: (unsigned)index withObject: (id)anObject;
+- (void) replaceObjectAtIndex: (NSUInteger)index withObject: (id)anObject;
 
 /**
  * Replaces the value of the named field.<br />
